Fixed bug so questions now appear
